requirements.txt
Django==5.0.6
psycopg2-binary==2.9.10
Faker==25.0.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-filter==24.2